% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidescroll.R
\name{is_sidescroller}
\alias{is_sidescroller}
\title{Check if is a side scroller object}
\usage{
is_sidescroller(x)
}
\arguments{
\item{x}{object to be checked if it is a side scroller}
}
\description{
Check if is a side scroller object
}
\examples{
x <- sidescroller()
is_sidescroller(x)
}
