% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_to_html.R
\name{markdown_to_html}
\alias{markdown_to_html}
\title{convert markdown text to html}
\usage{
markdown_to_html(x)
}
\description{
convert markdown text to html
}
\details{
A tool for converting passed text in RMD format into html for a slide
}
