% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{panel}
\alias{panel}
\title{Contain the contents of a panel}
\usage{
panel(..., style = NULL)
}
\arguments{
\item{...}{contents of the panel}

\item{style}{panel css styling}
}
\description{
helper function for providing contents for multi-panel slides
}
