% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{panel_overlay}
\alias{panel_overlay}
\title{Overlay the contents of a panel over another panel prior}
\usage{
panel_overlay(x, style, overlay = 50)
}
\arguments{
\item{x}{character vector of contents written in markdown}

\item{style}{panel css styling}

\item{overlay}{integer value between 1 and 100. Represents the percent overlay of prior panel.}
}
\description{
Overlay the contents of a panel over the prior panel, Useful for items "appearing" in place.
}
