% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sidescroller}
\alias{save_sidescroller}
\title{Save the sidescroller object}
\usage{
save_sidescroller(x, file, background = "#313131", libdir = "lib")
}
\arguments{
\item{file}{File to write content to}

\item{background}{Background color for web page}

\item{libdir}{Directory to copy dependenies to}
}
\description{
Utility to save the sidescroller object
}
