% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidescroll.R
\name{sidescroller}
\alias{sidescroller}
\title{Side Scroller Initializer}
\usage{
sidescroller(title = "", author = "", ...)
}
\arguments{
\item{title}{define the title of the presentation}

\item{author}{define the auther of the presentation}

\item{...}{HTML tags to add to head}
}
\description{
Side Scroller Initializer
}
\details{
initialize a side scroller object
}
