% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_ext.R
\name{slide}
\alias{slide}
\title{Basic Slide with title and contents}
\usage{
slide(.x, title = NULL, ...)
}
\arguments{
\item{.x}{a sidescoller object}

\item{title}{pass a title for the slide}

\item{...}{html content to create slide}
}
\description{
basic slide with title and contents
}
\examples{
sidescroller() \%>\% 
    slide(
    title = "Markdown Example Slide",
    p("This content will be rendered")
    p("as html content.")
    )
}
