% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_base}
\alias{slide_base}
\title{Generic Slide Creation}
\usage{
slide_base(x, ..., background = NULL, visibility = FALSE,
  slide_id = "generic", slide_class = "generic")
}
\arguments{
\item{x}{a sidescroller object}

\item{...}{tag elements to be added to the slide}

\item{background}{What color should the background of the slide be}

\item{visibility}{Should the slide default to be visible before sliding over. Defaults to false}

\item{slide_id}{set the css slide id}

\item{slide_class}{set the css slide class}
}
\description{
helper function for creating slides. Intended for use within other functions
}
\examples{
x <- sidescroller() \%>\% 
    slide_base(h1("This is a new slide),br(),p("This is some slide contents"), background = "#111111")

}
