% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_ext.R
\name{slide_wide_markdown}
\alias{slide_wide_markdown}
\title{Slide made to fill entire screen}
\usage{
slide_wide_markdown(.x, title = NULL, x)
}
\arguments{
\item{.x}{a side scroller object}

\item{title}{title of the slide}

\item{x}{slide contents written in markdown}
}
\description{
Make a full width slide
}
\examples{
sidescroller() \%>\% 
   slide_wide_markdown(
     title = "This is a wide slide based on markdown",
     "This slide will fill the width of the display.
      The purpose could be to clear out the past display, or stage before the next section.
      The Conents is written in _markdown_.")
     )
}
