% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_slide.R
\name{title_slide}
\alias{title_slide}
\title{Create title slide for presentation}
\usage{
title_slide(x, title = NULL, subtitle = NULL, ...,
  text_align = c("center", "left", "right"))
}
\arguments{
\item{x}{a side scoller object}

\item{title}{Title of the presentation}

\item{subtitle}{a character vector of subtitles to add}

\item{...}{arguments to be passed to \code{\link[sidescroller]{slide_base}}}

\item{text_align}{Where should the title and subtitles be aligned? center, left or right.}
}
\description{
Create title slide for presentation
}
\details{
create a title slide for your presentation
}
\examples{
sidescroller() \%>\% 
   title_slide(
     title = "Sidescoller Presentation", 
     subtitle = "By Ellis Hughes")
}
