% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidescroll.R
\name{sidescroller}
\alias{sidescroller}
\title{Side Scroller Initializer
Initialize a Side Scroller}
\usage{
sidescroller()
}
\description{
initialize a side scroller object
}
\examples{
x <- sidescroller()
}
