% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidescroll.R
\name{sidescroller}
\alias{sidescroller}
\title{Side Scroller Initializer
Initialize a Side Scroller}
\usage{
sidescroller(html_dependencies = NULL)
}
\arguments{
\item{html_dependencies}{list of html dependencies, generated by \code{htmltools::htmlDependency}.}
}
\description{
initialize a side scroller object
}
\examples{
x <- sidescroller()

## Add font-awesome dependency
x <- sidescroller(
         list(htmltools::htmlDependency(
            name = "font-awesome",
            version = "4.7.0",
            src = c(href = "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/"),
            stylesheet = "font-awesome.min.css")
         )))

}
