% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_base}
\alias{slide_base}
\title{Generic Slide Creation}
\usage{
slide_base(x, ..., style = NULL, visibility = FALSE, slide_id = NULL,
  slide_class = NULL, container_class = "fit_content")
}
\arguments{
\item{x}{a sidescroller object}

\item{...}{tag elements to be added to the slide}

\item{style}{style to be passed to the slide}

\item{visibility}{Should the slide default to be visible before sliding over. Defaults to false}

\item{slide_id}{set the slide id for css}

\item{slide_class}{set the slide class for css}

\item{container_class}{set the class slide container class for css}
}
\description{
helper function for creating slides. Intended for use within other functions
}
\examples{
x <- sidescroller() \%>\% 
    slide_base(
    h1("This is a new slide),
    br(),
    p("This is some slide contents"),
    style = "background:#111111;")

}
