% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_ext.R
\name{slide_wide}
\alias{slide_wide}
\title{Slide made to fill entire screen}
\usage{
slide_wide(x, ..., title = NULL)
}
\arguments{
\item{x}{a side scroller object}

\item{...}{slide contents and arguments to be passed to \code{\link[sidescroller]{slide_base}}}

\item{title}{title of the slide}
}
\description{
Make a full width slide
}
\examples{
sidescroller() \%>\% 
   slide_wide(
     title = "This is a wide slide",
     p("This slide will fill the width of the display"),
     p("The purpose could be to clear out the past display, or stage before the next section")
     )
}
